ACS-6389: add new common-test mvn module, refactor and clean-up code
