+++ bulk-ingester/src/integration-test/java/org/alfresco/hxi_connector/bulk_ingester/util/integration/PostgresIntegrationTestBase.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -35,7 +35,7 @@ import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 
-import org.alfresco.hxi_connector.bulk_ingester.util.DockerTags;
+import org.alfresco.hxi_connector.common.test.util.DockerTags;
 
 @Testcontainers
 @SuppressWarnings({"PMD.UseUtilityClass", "PMD.UnusedPrivateMethod"})
