+++ common-test/src/main/resources/docker-tags.properties
@@ -1,3 +1,4 @@
+postgres.tag=@dist.postgres.version@
+activemq.tag=@dist.activemq.docker.tag@
 localstack.tag=@dist.localstack.version@
 wiremock.tag=@dist.wiremock.version@
-activemq.tag=@dist.activemq.docker.tag@
