+++ live-ingester/src/integration-test/java/org/alfresco/hxi_connector/live_ingester/adapters/messaging/hx_insight/HxInsightEventPublisherIntegrationTest.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -38,7 +38,7 @@ import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.times;
 
-import static org.alfresco.hxi_connector.live_ingester.util.AuthUtils.AUTH_HEADER;
+import static org.alfresco.hxi_connector.live_ingester.util.auth.AuthUtils.AUTH_HEADER;
 
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.matching.ContainsPattern;
@@ -61,6 +61,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 import org.wiremock.integrations.testcontainers.WireMockContainer;
 
+import org.alfresco.hxi_connector.common.test.util.DockerTags;
 import org.alfresco.hxi_connector.live_ingester.adapters.config.IntegrationProperties;
 import org.alfresco.hxi_connector.live_ingester.domain.exception.EndpointClientErrorException;
 import org.alfresco.hxi_connector.live_ingester.domain.exception.EndpointServerErrorException;
@@ -68,9 +69,8 @@ import org.alfresco.hxi_connector.live_ingester.domain.ports.ingestion_engine.In
 import org.alfresco.hxi_connector.live_ingester.domain.ports.ingestion_engine.NodeEvent;
 import org.alfresco.hxi_connector.live_ingester.domain.ports.ingestion_engine.UpdateNodeMetadataEvent;
 import org.alfresco.hxi_connector.live_ingester.domain.usecase.metadata.model.EventType;
-import org.alfresco.hxi_connector.live_ingester.util.DockerTags;
-import org.alfresco.hxi_connector.live_ingester.util.WithMockOAuth2User;
-import org.alfresco.hxi_connector.live_ingester.util.WithoutAnyUser;
+import org.alfresco.hxi_connector.live_ingester.util.auth.WithMockOAuth2User;
+import org.alfresco.hxi_connector.live_ingester.util.auth.WithoutAnyUser;
 
 @SpringBootTest(classes = {
         IntegrationProperties.class,
