+++ live-ingester/src/integration-test/java/org/alfresco/hxi_connector/live_ingester/adapters/messaging/hx_insight/storage/connector/PreSignedUrlRequesterIntegrationTest.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -40,7 +40,7 @@ import static org.mockito.BDDMockito.then;
 import static org.mockito.Mockito.times;
 
 import static org.alfresco.hxi_connector.live_ingester.adapters.messaging.hx_insight.storage.connector.PreSignedUrlRequester.STORAGE_LOCATION_PROPERTY;
-import static org.alfresco.hxi_connector.live_ingester.util.AuthUtils.AUTH_HEADER;
+import static org.alfresco.hxi_connector.live_ingester.util.auth.AuthUtils.AUTH_HEADER;
 
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -70,12 +70,12 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 import org.wiremock.integrations.testcontainers.WireMockContainer;
 
+import org.alfresco.hxi_connector.common.test.util.DockerTags;
 import org.alfresco.hxi_connector.live_ingester.adapters.config.IntegrationProperties;
 import org.alfresco.hxi_connector.live_ingester.domain.exception.EndpointClientErrorException;
 import org.alfresco.hxi_connector.live_ingester.domain.exception.EndpointServerErrorException;
-import org.alfresco.hxi_connector.live_ingester.util.DockerTags;
-import org.alfresco.hxi_connector.live_ingester.util.WithMockOAuth2User;
-import org.alfresco.hxi_connector.live_ingester.util.WithoutAnyUser;
+import org.alfresco.hxi_connector.live_ingester.util.auth.WithMockOAuth2User;
+import org.alfresco.hxi_connector.live_ingester.util.auth.WithoutAnyUser;
 
 @SpringBootTest(classes = {
         IntegrationProperties.class,
