--- live-ingester/src/integration-test/java/org/alfresco/hxi_connector/live_ingester/util/DockerTagsIntegrationTest.java
@@ -1,47 +1,0 @@
-/*
- * #%L
- * Alfresco HX Insight Connector
- * %%
- * Copyright (C) 2024 Alfresco Software Limited
- * %%
- * This file is part of the Alfresco software.
- * If the software was purchased under a paid Alfresco license, the terms of
- * the paid license agreement will prevail.  Otherwise, the software is
- * provided under the following open source license terms:
- *
- * Alfresco is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * Alfresco is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
- * #L%
- */
-package org.alfresco.hxi_connector.live_ingester.util;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.catchThrowable;
-
-import org.junit.jupiter.api.Test;
-
-class DockerTagsIntegrationTest
-{
-
-    @Test
-    void testMavenPropertyResolution()
-    {
-        // when
-        Throwable thrown = catchThrowable(() -> DockerTags.keySet().stream()
-                .map(String::valueOf)
-                .forEach(DockerTags::getProperty));
-
-        // then
-        assertThat(thrown).doesNotThrowAnyException();
-    }
-}
