+++ live-ingester/src/integration-test/java/org/alfresco/hxi_connector/live_ingester/util/auth/WithMockOAuth2UserSecurityContextFactory.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -23,7 +23,7 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.hxi_connector.live_ingester.util;
+package org.alfresco.hxi_connector.live_ingester.util.auth;
 
 import static org.alfresco.hxi_connector.live_ingester.adapters.auth.AuthenticationService.ENVIRONMENT_KEY_ATTRIBUTE_KEY;
