+++ live-ingester/src/integration-test/java/org/alfresco/hxi_connector/live_ingester/util/auth/WithoutAnyUser.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -23,7 +23,7 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.hxi_connector.live_ingester.util;
+package org.alfresco.hxi_connector.live_ingester.util.auth;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -31,6 +31,6 @@ import java.lang.annotation.RetentionPolicy;
 import org.springframework.security.test.context.support.WithSecurityContext;
 
 @Retention(RetentionPolicy.RUNTIME)
-@WithSecurityContext(factory = WithMockOAuth2UserSecurityContextFactory.class)
-public @interface WithMockOAuth2User
+@WithSecurityContext(factory = WithoutAnyUserSecurityContextFactory.class)
+public @interface WithoutAnyUser
 {}
