+++ live-ingester/src/integration-test/java/org/alfresco/hxi_connector/live_ingester/util/auth/WithoutAnyUserSecurityContextFactory.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -23,14 +23,17 @@
  * along with Alfresco. If not, see <http://www.gnu.org/licenses/>.
  * #L%
  */
-package org.alfresco.hxi_connector.live_ingester.util;
+package org.alfresco.hxi_connector.live_ingester.util.auth;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.test.context.support.WithSecurityContextFactory;
 
-import org.springframework.security.test.context.support.WithSecurityContext;
-
-@Retention(RetentionPolicy.RUNTIME)
-@WithSecurityContext(factory = WithoutAnyUserSecurityContextFactory.class)
-public @interface WithoutAnyUser
-{}
+public class WithoutAnyUserSecurityContextFactory implements WithSecurityContextFactory<WithoutAnyUser>
+{
+    @Override
+    public SecurityContext createSecurityContext(WithoutAnyUser annotation)
+    {
+        return SecurityContextHolder.createEmptyContext();
+    }
+}
