+++ pom.xml
@@ -14,6 +14,7 @@
     <packaging>pom</packaging>
 
     <modules>
+        <module>common-test</module>
         <module>bulk-ingester</module>
         <module>live-ingester</module>
         <module>prediction-applier</module>
@@ -59,9 +60,10 @@
         <skipUnitTests>${skipTests}</skipUnitTests>
         <skipIntegrationTests>${skipTests}</skipIntegrationTests>
 
-        <alfresco.platform.version>23.1.0</alfresco.platform.version>
-        <alfresco.sdk.version>4.7.0</alfresco.sdk.version>
-        <acs-event-model.version>0.0.27</acs-event-model.version>
+        <alfresco-platform.version>23.1.0</alfresco-platform.version>
+        <alfresco-sdk.version>4.7.0</alfresco-sdk.version>
+        <alfresco-db-connector.version>4.0.1</alfresco-db-connector.version>
+        <alfresco-event-model.version>0.0.27</alfresco-event-model.version>
         <spring-boot.version>3.2.3</spring-boot.version>
         <spring-camel.version>4.4.0</spring-camel.version>
         <commons-lang.version>3.14.0</commons-lang.version>
@@ -75,15 +77,17 @@
 
         <maven-resources-plugin.version>3.3.1</maven-resources-plugin.version>
         <maven-assembly-plugin.version>3.6.0</maven-assembly-plugin.version>
+        <maven-jar-plugin.version>3.3.0</maven-jar-plugin.version>
         <license-maven-plugin.version>2.3.0</license-maven-plugin.version>
         <jacoco-maven-plugin.version>0.8.11</jacoco-maven-plugin.version>
         <spotless-plugin.version>2.43.0</spotless-plugin.version>
+        <jrebel-maven-plugin.version>1.2.0</jrebel-maven-plugin.version>
 
         <!-- Used by the generated .env file and Dockerfile for docker-compose -->
         <dist.alfresco.docker.user>alfresco</dist.alfresco.docker.user>
         <dist.alfresco.docker.image>quay.io/alfresco/alfresco-content-repository</dist.alfresco.docker.image>
-        <dist.alfresco.version>${alfresco.platform.version}</dist.alfresco.version>
-        <dist.share.version>${alfresco.platform.version}</dist.share.version>
+        <dist.alfresco.version>${alfresco-platform.version}</dist.alfresco.version>
+        <dist.share.version>${alfresco-platform.version}</dist.share.version>
         <dist.postgres.version>14.4</dist.postgres.version>
         <dist.transform.router.version>4.0.1</dist.transform.router.version>
         <dist.transform.core.aio.version>5.0.1</dist.transform.core.aio.version>
@@ -95,7 +99,7 @@
         <dist.nginx.version>3.4.2</dist.nginx.version>
         <dist.elasticsearch.version>7.17.15</dist.elasticsearch.version>
         <dist.search.enterprise.version>4.0.0.1</dist.search.enterprise.version>
-        <dist.localstack.version>3.0.2</dist.localstack.version>
+        <dist.localstack.version>3.2.0</dist.localstack.version>
         <dist.wiremock.version>${wiremock.version}</dist.wiremock.version>
     </properties>
 
@@ -169,9 +173,30 @@
 
             <dependency>
                 <groupId>org.alfresco</groupId>
-                <artifactId>acs-event-model</artifactId>
-                <version>${acs-event-model.version}</version>
+                <artifactId>alfresco-hxinsight-connector-common-test</artifactId>
+                <version>${project.version}</version>
             </dependency>
+            <dependency>
+                <groupId>org.alfresco</groupId>
+                <artifactId>alfresco-elasticsearch-database-connector</artifactId>
+                <version>${alfresco-db-connector.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.alfresco</groupId>
+                <artifactId>alfresco-elasticsearch-database-connector-api</artifactId>
+                <version>${alfresco-db-connector.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.alfresco</groupId>
+                <artifactId>acs-event-model</artifactId>
+                <version>${alfresco-event-model.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.alfresco.maven</groupId>
+                <artifactId>alfresco-rad</artifactId>
+                <version>${alfresco-sdk.version}</version>
+            </dependency>
+
             <dependency>
                 <groupId>org.apache.commons</groupId>
                 <artifactId>commons-lang3</artifactId>
@@ -253,6 +278,11 @@
                 <plugin>
                     <artifactId>maven-assembly-plugin</artifactId>
                     <version>${maven-assembly-plugin.version}</version>
+                </plugin>
+
+                <plugin>
+                    <artifactId>maven-jar-plugin</artifactId>
+                    <version>${maven-jar-plugin.version}</version>
                 </plugin>
 
                 <plugin>
