+++ prediction-applier/src/integration-test/java/org/alfresco/hxi_connector/prediction_applier/hxinsight/PredictionListenerIntegrationTest.java
@@ -2,7 +2,7 @@
  * #%L
  * Alfresco HX Insight Connector
  * %%
- * Copyright (C) 2024 Alfresco Software Limited
+ * Copyright (C) 2023 - 2024 Alfresco Software Limited
  * %%
  * This file is part of the Alfresco software.
  * If the software was purchased under a paid Alfresco license, the terms of
@@ -49,9 +49,9 @@ import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.utility.DockerImageName;
 
-import org.alfresco.hxi_connector.prediction_applier.util.DockerTags;
-import org.alfresco.hxi_connector.prediction_applier.util.LoggingUtils;
-import org.alfresco.hxi_connector.prediction_applier.util.RetryUtils;
+import org.alfresco.hxi_connector.common.test.util.DockerTags;
+import org.alfresco.hxi_connector.common.test.util.LoggingUtils;
+import org.alfresco.hxi_connector.common.test.util.RetryUtils;
 import org.alfresco.hxi_connector.prediction_applier.util.local.LocalSqsPublisher;
 
 @SpringBootTest(classes = {
