--- prediction-applier/src/integration-test/resources/docker-tags.properties
@@ -1,1 +1,0 @@
-localstack.tag=@dist.localstack.version@
